fix - change z-index dialog
this fix is to make sure the dialog always on the top of all element.

BREAKING CHANGE:
before this fix, button for changing theme positioned on the top of dialog (z-index: 10), this issue is visible on mobile device. Behavior change from the overlapping changing theme button to dialog always on top of element